refactor middleware types into its own directory, implement action payload separation in onIncomingEvent and builtin middleware
there is still one problem that remains with the type defintion of
SlackAction - it isn't yet useful for defining custom BlockAction
types. the details are spelled out in the comments. this isn't a
regression because it wouldn't have worked before.
